Move xssSantizer and tests to utils, update imports
Moved xssSantizer to utils, also moved its tests to utils. Updated all
imports of xssSantizer to use the new path.
